support-agent/pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[bey,openai]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
